[JobController] Separated to several controllers.
