+++ .idea/dataSources/6d93cf7f-dcf3-464a-a150-a540bee9dcf7/storage_v2/_src_/database/postgres.edMnLQ/schema/information_schema.FNRwLQ.meta
@@ -1,0 +1,2 @@
+#n:information_schema
+!<md> [null, 0, null, null, -2147483648, -2147483648]
