+++ .idea/dataSources/6d93cf7f-dcf3-464a-a150-a540bee9dcf7/storage_v2/_src_/database/postgres.edMnLQ/schema/public.abK9xQ.meta
@@ -1,0 +1,2 @@
+#n:public
+!<md> [5913340, 0, null, null, -2147483648, -2147483648]
