+++ .idea/sqldialects.xml
@@ -2,5 +2,6 @@
 <project version="4">
   <component name="SqlDialectMappings">
     <file url="file://$PROJECT_DIR$/src/controllers/JobController.ts" dialect="PostgreSQL" />
+    <file url="file://$PROJECT_DIR$/src/controllers/TasksController.ts" dialect="PostgreSQL" />
   </component>
 </project>
